+++ Bot/Libs/genshin_wish_sim_utils/wish.py
@@ -1,10 +1,9 @@
 import asyncio
-import random
 
 import numpy as np
 import uvloop
 from numpy.random import default_rng
-from sqlalchemy import select, update
+from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
 
@@ -26,129 +25,6 @@ class KumikoWSUtils:
         engine = create_async_engine(uri)
         async with engine.begin() as conn:
             await conn.run_sync(Base.metadata.create_all)
-
-    async def addToWSData(
-        self,
-        uuid: str,
-        event_name: str,
-        name: str,
-        description: str,
-        star_rank: int,
-        type: str,
-        uri: str,
-    ) -> None:
-        """Adds an entry into the WS DB
-
-        Args:
-            uuid (str): WS Item UUID
-            event_name (str): The name of the event
-            name (str): Name of the item or character
-            description (str): Description of the item or character
-            star_rank (int): The rank of the item or character
-            type (str): Character or Item?
-            uri (str): Connection URI
-        """
-        engine = create_async_engine(uri)
-        asyncSession = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-        async with asyncSession() as session:
-            async with session.begin():
-                insertItem = models.WSData(
-                    uuid=uuid,
-                    event_name=event_name,
-                    name=name,
-                    description=description,
-                    star_rank=star_rank,
-                    type=type,
-                )
-                session.add_all([insertItem])
-                await session.commit()
-
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-
-    async def selectAllWSData(self, uri: str) -> list:
-        """Selects literally all of the data out of the DB.
-        Only for testing purposes
-
-        Args:
-            uri (str): Connection URI
-
-        Returns:
-            list: List of results
-        """
-        engine = create_async_engine(uri)
-        asyncSession = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-        async with asyncSession() as session:
-            async with session.begin():
-                selectItem = select(models.WSData)
-                res = await session.execute(selectItem)
-                return np.array([row for row in res.scalars()])
-
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-
-    async def updateWSDataDescription(
-        self, uuid: str, description: str, uri: str
-    ) -> None:
-        """Updates the WS Item with a different description
-
-        Args:
-            uuid (str): WS Item UUID
-            description (str): New Description
-            uri (str): Connection URI
-        """
-        engine = create_async_engine(uri)
-        asyncSession = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-        async with asyncSession() as session:
-            async with session.begin():
-                updateItem = update(
-                    models.WSData, values={models.WSData.description: description}
-                ).where(models.WSData.uuid == uuid)
-                await session.execute(updateItem)
-                await session.commit()
-
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-
-    async def getWSItemName(self, name: str, uri: str) -> list:
-        """Returns an item from the WS DB based on that name
-
-        Args:
-            name (str): WS Item Name
-            uri (str): Connection URI
-
-        Returns:
-            list: List of items
-        """
-        engine = create_async_engine(uri)
-        asyncSession = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-        async with asyncSession() as session:
-            async with session.begin():
-                selItem = select(models.WSData).where(models.WSData.name == name)
-                res = await session.execute(selItem)
-                return [row for row in res.scalars()]
-
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-
-    async def getRandomWS(self, uri: str) -> list:
-        """Randomly picks out a row based on the WS Dataset
-
-        Args:
-            uri (str): Connection URI
-
-        Returns:
-            list: A list of results
-        """
-        engine = create_async_engine(uri)
-        asyncSession = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-        async with asyncSession() as session:
-            async with session.begin():
-                selectItem = (
-                    select(models.WSData)
-                    .offset(random.randrange(25))  # nosec B311
-                    .limit(1)
-                )
-                res = await session.execute(selectItem)
-                return [row for row in res.scalars()]
-
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
     async def getRandomWSArray(self, star_rank: int, uri: str) -> models.WSData:
         """Selects all of the data from the WS DB, and randomly picks out a row.